Extract rollbar configurator and only auto use if config defined (rails)
